refactor(runner): move `findAndInlineDataURILinks` into `data-uri.ts`, making the dependency one-way
The colocation left a new import cycle: `data-uri.ts` needs the link
helpers from `link-utils.ts`, while `link-utils.ts` imported
`getJSONFromDataURI()` back -- solely for `findAndInlineDataURILinks`.
That function is `data:` URI logic (it dissolves `data:` URI links back
into the values they carry, as a direct consumer of the decode half), so
it moves here, and the dependency now runs one way only: `data-uri.ts`
depends on the general link machinery, never the reverse.

Its two focused tests move from `link-utils.test.ts` into the
already-dedicated `data-uri-inlining.test.ts`; all importers rewired.
