+++ packages/runner/src/cfc/ui-contract.ts
@@ -1,12 +1,8 @@
 import { isRecord } from "@commonfabric/utils/types";
 import type { CellScope, JSONSchema } from "../builder/types.ts";
-import {
-  findAndInlineDataURILinks,
-  type NormalizedFullLink,
-  parseLink,
-} from "../link-utils.ts";
+import { type NormalizedFullLink, parseLink } from "../link-utils.ts";
 import type { IExtendedStorageTransaction } from "../storage/interface.ts";
-import { getJSONFromDataURI } from "../data-uri.ts";
+import { findAndInlineDataURILinks, getJSONFromDataURI } from "../data-uri.ts";
 import { ContextualFlowControl } from "../cfc.ts";
 import type { CfcAddress } from "./types.ts";
 import { isNormalizedFullLink } from "../link-types.ts";
